+++ examples/testing/lib/main.dart
@@ -56,12 +56,14 @@ class _HomeState extends State<Home> {
           onEventChanged: (event, updatedEvent) =>
               eventsController.updateEvent(event: event, updatedEvent: updatedEvent),
         ),
-        header: KalenderHeader(multiDayTileComponents: _multiDayTileComponents),
-        body: KalenderBody(
-          multiDayTileComponents: _tileComponents,
-          monthTileComponents: _multiDayTileComponents,
-          scheduleTileComponents: _scheduleTileComponents,
-        ),
+        views: [
+          MultiDayViewParts(
+            header: MultiDayHeader(tileComponents: _multiDayTileComponents),
+            body: MultiDayBody(tileComponents: _tileComponents),
+          ),
+          MonthViewParts(body: MonthBody(tileComponents: _multiDayTileComponents)),
+          ScheduleViewParts(body: ScheduleBody(tileComponents: _scheduleTileComponents)),
+        ],
       ),
     );
   }