+++ lib/kalender.dart
@@ -8,10 +8,8 @@
 
 library;
 
-export 'package:kalender/src/kalender_body.dart';
-export 'package:kalender/src/kalender_header.dart';
-
 export 'package:kalender/src/kalender_view.dart';
+export 'package:kalender/src/view_parts.dart';
 
 export 'package:kalender/src/enumerations.dart';
 
@@ -73,7 +71,6 @@ export 'package:kalender/src/widgets/month/month_header.dart';
 export 'package:kalender/src/widgets/multi_day/multi_day_body.dart';
 export 'package:kalender/src/widgets/multi_day/multi_day_header.dart';
 export 'package:kalender/src/widgets/schedule/schedule_body.dart';
-export 'package:kalender/src/widgets/schedule/schedule_header.dart';
 
 export 'package:kalender/src/models/mixins/event_tile_utils.dart';
 export 'package:kalender/src/models/mixins/drag_target_utils.dart';