+++ lib/src/widgets/multi_day/multi_day_body.dart
@@ -12,6 +12,7 @@ import 'package:kalender/src/widgets/draggable/day_draggable.dart';
 import 'package:kalender/src/widgets/events_widgets/day_events_widget.dart';
 import 'package:kalender/src/widgets/internal_components/time_indicator_positioner.dart';
 import 'package:kalender/src/widgets/internal_components/timeline_sizer.dart';
+import 'package:kalender/src/widgets/internal_components/view_providers.dart';
 import 'package:linked_pageview/linked_pageview.dart';
 
 // TODO: I want to simplify this widget by removing the split between the content and header, essentially removing the duplicate page views,
@@ -28,13 +29,55 @@ class MultiDayBody extends StatelessWidget {
   /// The [MultiDayBodyConfiguration] that will be used by the [MultiDayBody].
   final MultiDayBodyConfiguration? configuration;
 
-  const MultiDayBody({super.key, this.configuration});
+  /// Overrides the [KalenderView.callbacks] for this widget.
+  final KalenderCallbacks? callbacks;
+
+  /// Overrides the [KalenderView.interaction] for this widget.
+  final KalenderInteraction? interaction;
+
+  /// The snapping of dragged and resized events. Defaults to [KalenderSnapping].
+  final KalenderSnapping? snapping;
+
+  /// The tile components. Defaults to [TileComponents.defaultComponents].
+  final TileComponents? tileComponents;
+
+  const MultiDayBody({
+    super.key,
+    this.configuration,
+    this.callbacks,
+    this.interaction,
+    this.snapping,
+    this.tileComponents,
+  });
 
   /// The key used to identify the [SingleChildScrollView] of the [MultiDayBody].
   static const singleChildScrollViewKey = ValueKey('singleChildScrollViewKey');
 
   /// The key used to identify the timeline gutter of the [MultiDayBody].
   static const timelineKey = ValueKey('MultiDayBody.timeline');
+
+  @override
+  Widget build(BuildContext context) {
+    assert(
+      context.viewController is MultiDayViewController,
+      'The view controller needs to be a $MultiDayViewController',
+    );
+    return ViewProviders(
+      callbacks: callbacks,
+      interaction: interaction,
+      snapping: snapping,
+      installSnapping: true,
+      tileComponents: tileComponents ?? TileComponents.defaultComponents(),
+      heightPerMinute: (context.viewController as MultiDayViewController).heightPerMinute,
+      child: _MultiDayBody(configuration: configuration),
+    );
+  }
+}
+
+class _MultiDayBody extends StatelessWidget {
+  final MultiDayBodyConfiguration? configuration;
+
+  const _MultiDayBody({this.configuration});
 
   @override
   Widget build(BuildContext context) {
@@ -61,7 +104,7 @@ class MultiDayBody extends StatelessWidget {
         Scrollbar(
           controller: viewController.scrollController,
           child: SingleChildScrollView(
-            key: singleChildScrollViewKey,
+            key: MultiDayBody.singleChildScrollViewKey,
             controller: viewController.scrollController,
             physics: configuration.scrollPhysics,
             child: SizedBox(
@@ -71,7 +114,7 @@ class MultiDayBody extends StatelessWidget {
                   // The timeline is always on the left side of the page, but should not scroll with the pageview.
                   // Its width is fixed to the shared timeline width so it aligns with the header and drag overlay.
                   SizedBox(
-                    key: timelineKey,
+                    key: MultiDayBody.timelineKey,
                     width: timelineWidth,
                     height: pageHeight,
                     child: bodyComponents.buildTimeline(