+++ test/widgets/header_month_builder_context_test.dart
@@ -29,8 +29,13 @@ void main() {
       eventsController: eventsController,
       kalenderController: KalenderController(viewConfiguration: viewConfiguration),
       components: components,
-      header: KalenderHeader(multiDayTileComponents: tiles),
-      body: KalenderBody(multiDayTileComponents: tiles),
+      views: [
+        MultiDayViewParts(
+          header: MultiDayHeader(tileComponents: tiles),
+          body: MultiDayBody(tileComponents: tiles),
+        ),
+        const MonthViewParts(),
+      ],
     );
     await pumpAndSettleWithMaterialApp(tester, theme == null ? view : KalenderTheme(data: theme, child: view));
   }