+++ test/widgets/page_boundary_test.dart
@@ -22,11 +22,7 @@ void main() {
     kalenderController = KalenderController(viewConfiguration: config);
     return pumpAndSettleWithMaterialApp(
       tester,
-      KalenderView(
-        eventsController: eventsController,
-        kalenderController: kalenderController,
-        body: const KalenderBody(),
-      ),
+      KalenderView(eventsController: eventsController, kalenderController: kalenderController, views: bodyOnlyViews),
     );
   }