+++ test/widgets/time_indicator_free_scroll_test.dart
@@ -50,8 +50,12 @@ void main() {
             ),
           ),
         ),
-        header: KalenderHeader(multiDayTileComponents: tiles),
-        body: KalenderBody(multiDayTileComponents: tiles),
+        views: [
+          MultiDayViewParts(
+            header: MultiDayHeader(tileComponents: tiles),
+            body: MultiDayBody(tileComponents: tiles),
+          ),
+        ],
       ),
     );
   }